+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Fixed
 
 - In the corpus selector, an empty folder would add 1 to the parent folder's corpus count
+- Searching by pressing Enter in Simple search is broken [#394](https://github.com/spraakbanken/korp-frontend/issues/394)
 
 ## [9.7.1] - 2024-09-18